+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-06-22
+### Changed
+- Use editor preference setting to store bookmarks instead of ScriptableObject
+- Swap bookmark and scene button to make it more intuitive
+- Now can only support Unity 6 since `SettingsProvider` starts being used
+
 ## [1.0.0] - 2022-01-05
 ### Added
 - First version of the package